Use dexie for sourcesDB
This commit is part of a db transition from lovefield to
any other database (in this case, dexie, which has had many
years of support). Lovefield has not been updated in 8 years,
and depends on packages which have critical vulnerabilities.

This commit removes nedb (which has not been used in many years),
and swaps in dexie for lovefield for RSS sources.

This commit creates the FluentDB database, which currently only
has one table: `sources`. This table will replace the original
sourcesDB. This attempts to migrate the original sourcesDB entries
seamlessly, then deletes the original DB.